pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "email-validator>=2.3.0",
    "fastapi>=0.122.0",
    "fastembed>=0.7.3",
    "google-generativeai>=0.8.5",
    "litellm>=1.80.7",
    "markdown-it-py>=4.0.0",
    "mdurl>=0.1.2",
    "openai>=2.8.1",
    "openai-agents[litellm]>=0.6.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "qdrant-client>=1.16.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]
